User Controller
created UserController and moved Register, show and update to from JWTController to it.
