+++ app/Http/Controllers/JWTController.php
@@ -22,30 +22,6 @@ class JWTController extends Controller
         $this->middleware(['jwt.verify','api'], ['except' => ['login', 'register']]);
     }
 
-    public function register(Request $request)
-    {
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|string|min:2|max:100',
-            'email' => 'required|string|email|max:100|unique:users',
-            'password' => 'required|string|confirmed|min:6',
-        ]);
-
-        if($validator->fails()) {
-            return response()->json($validator->errors(), 400);
-        }
-
-        $user = User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password)
-        ]);
-
-        return response()->json([
-            'message' => 'User successfully registered',
-            'user' => new UserResource($user),
-        ], 201);
-    }
-
     public function login(Request $request)
     {
         $validator = Validator::make($request->all(), [
@@ -77,11 +53,6 @@ class JWTController extends Controller
         return $this->respondWithToken(auth()->refresh());
     }
 
-    public function profile()
-    {
-        return response()->json(new UserResource(auth()->user()));
-    }
-
     protected function respondWithToken($token)
     {
         return response()->json([
@@ -91,28 +62,4 @@ class JWTController extends Controller
         ]);
     }
 
-    public function update(Request $request)
-    {
-        $user = auth()->user();
-        $validator = Validator::make($request->all(), [
-            'name' => 'required|string|min:2|max:100',
-            'email' => 'required|string|email|max:100|unique:users,email,'.$user->id,
-            'password' => 'required|string|confirmed|min:6',
-        ]);
-
-        if($validator->fails()) {
-            return response()->json($validator->errors(), 400);
-        }
-
-        $user->update([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password)
-        ]);
-
-        return response()->json([
-            'message' => 'User successfully updated',
-            'user' => new UserResource($user),
-        ], 201);
-    }
 }
